third_party/WebKit/LayoutTests/fast/dom/event-target-arguments.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<body>
<script>
description('EventTarget.addEventListener and EventTarget.removeEventLister should throw on invalid arguments.');

function listener(event)
{
}

debug('Signature:')
debug('void addEventListener(DOMString type, EventListener listener, optional boolean useCapture)');
shouldThrow('document.addEventListener()');
shouldThrow('document.addEventListener("foo")');
shouldNotThrow('document.addEventListener("foo", listener)');
shouldNotThrow('document.addEventListener("", listener)');
shouldNotThrow('document.addEventListener("", function(){})');
shouldNotThrow('document.addEventListener("bar", listener, false)');
shouldNotThrow('document.addEventListener("bar", listener, true)');
// null
shouldThrow('document.addEventListener(null)');
shouldNotThrow('document.addEventListener(null, listener)'); // converted to "null"
shouldBe('document.addEventListener("foo", null)', 'undefined');
shouldNotThrow('document.addEventListener("zork", listener, null)');
// undefined
shouldThrow('document.addEventListener(undefined)');
shouldNotThrow('document.addEventListener(undefined, listener)', 'undefined'); // converted to "undefined"
shouldBe('document.addEventListener("foo", undefined)', 'undefined');
shouldNotThrow('document.addEventListener("zork", listener, undefined)');

debug('');
debug('Signature:');
debug('void removeEventListener(DOMString type, EventListener listener, optional boolean useCapture)');
shouldThrow('document.removeEventListener()');
shouldThrow('document.removeEventListener("foo")');
shouldNotThrow('document.removeEventListener("foo", listener)');
shouldNotThrow('document.removeEventListener("foo", listener, true)');
shouldNotThrow('document.removeEventListener("bar", listener, false)');
shouldNotThrow('document.removeEventListener("bar", listener, false)');
shouldNotThrow('document.removeEventListener("bar", listener, true)');
// null
shouldThrow('document.removeEventListener(null)');
shouldNotThrow('document.removeEventListener(null, listener)'); // converted to "null"
shouldBe('document.removeEventListener("foo", null)', 'undefined');
shouldNotThrow('document.removeEventListener("zork", listener, null)');
// undefined
shouldThrow('document.removeEventListener(undefined)');
shouldNotThrow('document.removeEventListener(undefined, listener)', 'undefined'); // converted to "undefined"
shouldBe('document.removeEventListener("foo", undefined)', 'undefined');
shouldNotThrow('document.removeEventListener("zork", listener, undefined)');
</script>
